+++ src/cluster/leiden/test.py
@@ -35,15 +35,6 @@ def mudata_custom_connectivities_key(input_data, random_h5mu_path):
     return output_path
 
 
-# @pytest.fixture
-# def random_h5mu_path(tmp_path):
-#     def wrapper():
-#         unique_filename = f"{str(uuid.uuid4())}.h5mu"
-#         temp_file = tmp_path / unique_filename
-#         return temp_file
-#     return wrapper
-
-
 @pytest.mark.parametrize("compression", ["gzip", ""])
 @pytest.mark.parametrize(
     "output_key,expected_output_key", [("fooleiden", "fooleiden"), ("", "leiden")]